manage api router, create audio.
